requirements.txt
pydub==0.25.1
pyannote.audio==3.3.1
torch==2.3.1
torchaudio==2.3.1
faster-whisper==1.0.2
groq==0.9.0
fpdf2==2.7.9
ffmpeg-python==0.2.0
python-dotenv==1.0.1
mutagen==1.47.0
tqdm==4.66.4
ttkbootstrap==1.10.1
opencv-python==4.10.0.84
numpy==1.26.3
sentence-transformers==3.0.1
scikit-learn==1.5.0